requirements.txt
camoufox[geoip]
python-dotenv
httpx
flask
